--- src/expenses/controllers/expenses.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ExpensesController } from './expenses.controller';
-
-describe('ExpensesController', () => {
-  let controller: ExpensesController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ExpensesController],
-    }).compile();
-
-    controller = module.get<ExpensesController>(ExpensesController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
